token_app/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
click==8.0.1
coverage==5.5
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
isort==5.9.3
mypy-extensions==0.4.3
pathspec==0.9.0
psycopg2-binary==2.9.1
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
regex==2021.8.3
sqlparse==0.4.1
tomli==1.2.1
